+++ HISTORY.rst
@@ -6,6 +6,7 @@ Changelog
 - Sentiment analyzers return namedtuples, e.g. ``Sentiment(polarity=0.12, subjectivity=0.34)``.
 - Memory usage improvements to NaiveBayesAnalyzer and basic_extractor (default feature extractor for classifiers module).
 - Add ``textblob.tokenizers.sent_tokenize`` and ``textblob.tokenizers.word_tokenize`` convenience functions.
+- Add ``textblob.classifiers.MaxEntClassifer``.
 
 0.8.2 (2013-12-21)
 ------------------